orchestrator: delete the dead per-chain conf code
