--- sidechain-orchestrator/config/thunder_conf.go
@@ -1,397 +1,0 @@
-package config
-
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"strings"
-	"sync"
-	"time"
-
-	"github.com/fsnotify/fsnotify"
-	"github.com/rs/zerolog"
-)
-
-const thunderConfFilename = "thunder.conf"
-
-// ThunderConfManager manages Thunder sidechain configuration.
-// 1:1 port of GenericSidechainConfProvider + ThunderConfProvider from Dart.
-type ThunderConfManager struct {
-	Config      *GenericAppConfig
-	ConfigPath  string
-	bitcoinConf *BitcoinConfManager
-	log         zerolog.Logger
-
-	// File watching
-	watcher   *fsnotify.Watcher
-	watchDone chan struct{}
-}
-
-// GenericAppConfig is a simple key-value config for sidechain apps.
-// Dart: GenericAppConfig class
-type GenericAppConfig struct {
-	Settings map[string]string
-}
-
-func NewGenericAppConfig() *GenericAppConfig {
-	return &GenericAppConfig{Settings: make(map[string]string)}
-}
-
-func ParseGenericAppConfig(content string) *GenericAppConfig {
-	c := NewGenericAppConfig()
-	for _, line := range strings.Split(content, "\n") {
-		trimmed := strings.TrimSpace(line)
-		if trimmed == "" || strings.HasPrefix(trimmed, "#") {
-			continue
-		}
-		if eqIdx := strings.Index(trimmed, "="); eqIdx > 0 {
-			key := strings.TrimSpace(trimmed[:eqIdx])
-			value := strings.TrimSpace(trimmed[eqIdx+1:])
-			c.Settings[key] = value
-		}
-	}
-	return c
-}
-
-func (c *GenericAppConfig) Serialize() string {
-	var b strings.Builder
-	for key, value := range c.Settings {
-		fmt.Fprintf(&b, "%s=%s\n", key, value)
-	}
-	return b.String()
-}
-
-func (c *GenericAppConfig) GetSetting(key string) string {
-	return c.Settings[key]
-}
-
-func (c *GenericAppConfig) SetSetting(key, value string) {
-	c.Settings[key] = value
-}
-
-func (c *GenericAppConfig) RemoveSetting(key string) {
-	delete(c.Settings, key)
-}
-
-// NewThunderConfManager creates a new ThunderConfManager.
-func NewThunderConfManager(bitcoinConf *BitcoinConfManager, log zerolog.Logger) (*ThunderConfManager, error) {
-	m := &ThunderConfManager{
-		bitcoinConf: bitcoinConf,
-		log:         log.With().Str("component", "thunder-conf").Logger(),
-	}
-	if err := m.LoadConfig(); err != nil {
-		return nil, fmt.Errorf("load thunder config: %w", err)
-	}
-	if err := m.SyncNetworkFromBitcoinConf(); err != nil {
-		m.log.Warn().Err(err).Msg("failed to sync thunder config from bitcoin conf")
-	}
-	return m, nil
-}
-
-// LoadConfig loads thunder.conf from file, or creates default if not exists.
-// Dart: GenericSidechainConfProvider.loadConfig (L120-146)
-func (m *ThunderConfManager) LoadConfig() error {
-	m.ConfigPath = m.getConfigPath()
-
-	data, err := os.ReadFile(m.ConfigPath)
-	if err == nil {
-		m.Config = ParseGenericAppConfig(string(data))
-		return nil
-	}
-
-	if !os.IsNotExist(err) {
-		return fmt.Errorf("read thunder config: %w", err)
-	}
-
-	// Create default config
-	content := m.GetDefaultConfig()
-	m.Config = ParseGenericAppConfig(content)
-
-	if mkErr := os.MkdirAll(filepath.Dir(m.ConfigPath), 0755); mkErr != nil {
-		m.log.Error().Err(mkErr).Msg("failed to create thunder config directory")
-	} else if wErr := os.WriteFile(m.ConfigPath, []byte(content), 0644); wErr != nil {
-		m.log.Error().Err(wErr).Str("path", m.ConfigPath).Msg("failed to write default thunder config")
-	} else {
-		m.log.Info().Str("path", m.ConfigPath).Msg("created default thunder config file")
-	}
-
-	return nil
-}
-
-// GetDefaultConfig generates the default thunder.conf content.
-// Dart: ThunderConfProvider.getDefaultConfig (L52-90)
-func (m *ThunderConfManager) GetDefaultConfig() string {
-	network := "signet"
-	if m.bitcoinConf != nil {
-		switch m.bitcoinConf.Network {
-		case NetworkSignet:
-			network = "signet"
-		case NetworkRegtest:
-			network = "regtest"
-		case NetworkECash:
-			network = "mainnet"
-		default:
-			network = "signet"
-		}
-	}
-
-	ports := getThunderNetworkPorts(network)
-
-	return fmt.Sprintf(`# Thunder Configuration - Generated by Thunder App
-# These settings are converted to CLI arguments when Thunder starts.
-
-# Run in headless mode (no GUI)
-headless=true
-
-# Log level for console output
-log-level=DEBUG
-
-# Log level for file output
-log-level-file=WARN
-
-# Network (signet or regtest)
-network=%s
-
-# RPC server address
-rpc-addr=%s
-
-# P2P networking address
-net-addr=%s
-
-# Mainchain (Enforcer) gRPC connection URL
-mainchain-grpc-url=%s
-`, network, ports["rpc-addr"], ports["net-addr"], ports["mainchain-grpc-url"])
-}
-
-// getThunderNetworkPorts returns network-specific port mappings.
-// Dart: ThunderConfProvider.getNetworkPorts (L27-48)
-func getThunderNetworkPorts(network string) map[string]string {
-	switch network {
-	case "regtest":
-		return map[string]string{
-			"rpc-addr":           "127.0.0.1:16009",
-			"net-addr":           "0.0.0.0:14009",
-			"mainchain-grpc-url": "http://localhost:50051",
-		}
-	case "mainnet":
-		return map[string]string{
-			"rpc-addr":           "127.0.0.1:26009",
-			"net-addr":           "0.0.0.0:24009",
-			"mainchain-grpc-url": "http://localhost:50051",
-		}
-	default: // signet
-		return map[string]string{
-			"rpc-addr":           "127.0.0.1:6009",
-			"net-addr":           "0.0.0.0:4009",
-			"mainchain-grpc-url": "http://localhost:50051",
-		}
-	}
-}
-
-// GetCliArgs converts current config to CLI args.
-// Dart: GenericSidechainConfProvider.getCliArgs (L176-198)
-// Skips 'network' key (Dart: ThunderConfProvider.skippedCliKeys L24)
-func (m *ThunderConfManager) GetCliArgs() []string {
-	var args []string
-	if m.Config == nil {
-		return args
-	}
-
-	for key, value := range m.Config.Settings {
-		// Dart L24: skippedCliKeys => ['network']
-		if key == "network" {
-			continue
-		}
-		switch value {
-		case "true":
-			args = append(args, fmt.Sprintf("--%s", key))
-		case "false":
-			continue
-		default:
-			if value != "" {
-				args = append(args, fmt.Sprintf("--%s=%s", key, value))
-			}
-		}
-	}
-
-	return args
-}
-
-// GetNetwork returns the current network from config.
-func (m *ThunderConfManager) GetNetwork() string {
-	if m.Config == nil {
-		return "signet"
-	}
-	n := m.Config.GetSetting("network")
-	if n == "" {
-		return "signet"
-	}
-	return n
-}
-
-// GetCurrentConfigContent returns config as string.
-// Dart: GenericSidechainConfProvider.getCurrentConfigContent (L149-154)
-func (m *ThunderConfManager) GetCurrentConfigContent() string {
-	if m.Config == nil {
-		return m.GetDefaultConfig()
-	}
-	return m.Config.Serialize()
-}
-
-// WriteConfig writes raw config content to file.
-// Dart: GenericSidechainConfProvider.writeConfig (L157-173)
-func (m *ThunderConfManager) WriteConfig(content string) error {
-	m.Config = ParseGenericAppConfig(content)
-
-	confPath := m.getConfigPath()
-	if err := os.MkdirAll(filepath.Dir(confPath), 0755); err != nil {
-		return fmt.Errorf("create dir: %w", err)
-	}
-	if err := os.WriteFile(confPath, []byte(content), 0644); err != nil {
-		return fmt.Errorf("write config: %w", err)
-	}
-
-	m.log.Info().Str("path", confPath).Msg("saved thunder config")
-	return nil
-}
-
-// SyncNetworkFromBitcoinConf syncs network setting from Bitcoin config.
-// Dart: GenericSidechainConfProvider.syncNetworkFromBitcoinConf (L78-101)
-func (m *ThunderConfManager) SyncNetworkFromBitcoinConf() error {
-	if m.Config == nil || m.bitcoinConf == nil {
-		return nil
-	}
-
-	var sidechainNetwork string
-	switch m.bitcoinConf.Network {
-	case NetworkRegtest:
-		sidechainNetwork = "regtest"
-	case NetworkECash:
-		sidechainNetwork = "mainnet"
-	default:
-		sidechainNetwork = "signet"
-	}
-
-	currentNetwork := m.Config.GetSetting("network")
-	if currentNetwork != sidechainNetwork {
-		m.Config.SetSetting("network", sidechainNetwork)
-		// Apply network-specific ports
-		ports := getThunderNetworkPorts(sidechainNetwork)
-		for key, value := range ports {
-			m.Config.SetSetting(key, value)
-		}
-		return m.saveConfig()
-	}
-
-	return nil
-}
-
-func (m *ThunderConfManager) saveConfig() error {
-	if m.Config == nil {
-		return nil
-	}
-	confPath := m.getConfigPath()
-	if err := os.MkdirAll(filepath.Dir(confPath), 0755); err != nil {
-		return err
-	}
-	if err := os.WriteFile(confPath, []byte(m.Config.Serialize()), 0644); err != nil {
-		return fmt.Errorf("save thunder config: %w", err)
-	}
-	m.log.Info().Str("path", confPath).Msg("saved thunder config")
-	return nil
-}
-
-func (m *ThunderConfManager) getConfigPath() string {
-	return filepath.Join(ThunderDirs.RootDir(), thunderConfFilename)
-}
-
-// StartWatching watches the thunder config directory for changes.
-func (m *ThunderConfManager) StartWatching() error {
-	watcher, err := fsnotify.NewWatcher()
-	if err != nil {
-		return fmt.Errorf("create watcher: %w", err)
-	}
-
-	confDir := filepath.Dir(m.getConfigPath())
-	if err := os.MkdirAll(confDir, 0755); err != nil {
-		_ = watcher.Close()
-		return fmt.Errorf("create watch dir: %w", err)
-	}
-
-	if err := watcher.Add(confDir); err != nil {
-		_ = watcher.Close()
-		return fmt.Errorf("watch dir: %w", err)
-	}
-
-	m.watcher = watcher
-	m.watchDone = make(chan struct{})
-
-	go m.watchLoop()
-
-	m.log.Debug().Str("dir", confDir).Msg("thunder config file watching enabled")
-	return nil
-}
-
-func (m *ThunderConfManager) StopWatching() {
-	if m.watcher != nil {
-		_ = m.watcher.Close()
-	}
-	if m.watchDone != nil {
-		<-m.watchDone
-	}
-}
-
-func (m *ThunderConfManager) watchLoop() {
-	defer close(m.watchDone)
-
-	var debounce *time.Timer
-	var mu sync.Mutex
-
-	for {
-		select {
-		case event, ok := <-m.watcher.Events:
-			if !ok {
-				return
-			}
-			if !strings.HasSuffix(event.Name, thunderConfFilename) {
-				continue
-			}
-			if event.Op&(fsnotify.Write|fsnotify.Create) == 0 {
-				continue
-			}
-
-			mu.Lock()
-			if debounce != nil {
-				debounce.Stop()
-			}
-			debounce = time.AfterFunc(500*time.Millisecond, func() {
-				m.reloadConfigFromFileSystem()
-			})
-			mu.Unlock()
-
-		case err, ok := <-m.watcher.Errors:
-			if !ok {
-				return
-			}
-			m.log.Error().Err(err).Msg("thunder config watcher error")
-		}
-	}
-}
-
-func (m *ThunderConfManager) reloadConfigFromFileSystem() {
-	m.log.Info().Msg("reloading thunder config due to file system change")
-
-	confPath := m.getConfigPath()
-	data, err := os.ReadFile(confPath)
-	if err != nil {
-		m.log.Error().Err(err).Msg("failed to read thunder config")
-		return
-	}
-
-	newConfig := ParseGenericAppConfig(string(data))
-	if m.Config != nil && m.Config.Serialize() == newConfig.Serialize() {
-		return
-	}
-
-	m.Config = newConfig
-}
