--- sidechain-orchestrator/config/zside_conf.go
@@ -1,341 +1,0 @@
-package config
-
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"strings"
-	"sync"
-	"time"
-
-	"github.com/fsnotify/fsnotify"
-	"github.com/rs/zerolog"
-)
-
-const zsideConfFilename = "zside.conf"
-
-// ZSideConfManager manages ZSide sidechain configuration.
-// 1:1 port of GenericSidechainConfProvider + ZSideConfProvider from Dart.
-type ZSideConfManager struct {
-	Config      *GenericAppConfig
-	ConfigPath  string
-	bitcoinConf *BitcoinConfManager
-	log         zerolog.Logger
-
-	// File watching
-	watcher   *fsnotify.Watcher
-	watchDone chan struct{}
-}
-
-// NewZSideConfManager creates a new ZSideConfManager.
-func NewZSideConfManager(bitcoinConf *BitcoinConfManager, log zerolog.Logger) (*ZSideConfManager, error) {
-	m := &ZSideConfManager{
-		bitcoinConf: bitcoinConf,
-		log:         log.With().Str("component", "zside-conf").Logger(),
-	}
-	if err := m.LoadConfig(); err != nil {
-		return nil, fmt.Errorf("load zside config: %w", err)
-	}
-	if err := m.SyncNetworkFromBitcoinConf(); err != nil {
-		m.log.Warn().Err(err).Msg("failed to sync zside config from bitcoin conf")
-	}
-	return m, nil
-}
-
-// LoadConfig loads zside.conf from file, or creates default if not exists.
-func (m *ZSideConfManager) LoadConfig() error {
-	m.ConfigPath = m.getConfigPath()
-
-	data, err := os.ReadFile(m.ConfigPath)
-	if err == nil {
-		m.Config = ParseGenericAppConfig(string(data))
-		return nil
-	}
-
-	if !os.IsNotExist(err) {
-		return fmt.Errorf("read zside config: %w", err)
-	}
-
-	// Create default config
-	content := m.GetDefaultConfig()
-	m.Config = ParseGenericAppConfig(content)
-
-	if mkErr := os.MkdirAll(filepath.Dir(m.ConfigPath), 0755); mkErr != nil {
-		m.log.Error().Err(mkErr).Msg("failed to create zside config directory")
-	} else if wErr := os.WriteFile(m.ConfigPath, []byte(content), 0644); wErr != nil {
-		m.log.Error().Err(wErr).Str("path", m.ConfigPath).Msg("failed to write default zside config")
-	} else {
-		m.log.Info().Str("path", m.ConfigPath).Msg("created default zside config file")
-	}
-
-	return nil
-}
-
-// GetDefaultConfig generates the default zside.conf content.
-func (m *ZSideConfManager) GetDefaultConfig() string {
-	network := "signet"
-	if m.bitcoinConf != nil {
-		switch m.bitcoinConf.Network {
-		case NetworkSignet:
-			network = "signet"
-		case NetworkRegtest:
-			network = "regtest"
-		case NetworkECash:
-			network = "mainnet"
-		default:
-			network = "signet"
-		}
-	}
-
-	ports := getZSideNetworkPorts(network)
-
-	return fmt.Sprintf(`# ZSide Configuration - Generated by ZSide App
-# These settings are converted to CLI arguments when ZSide starts.
-
-# Run in headless mode (no GUI)
-headless=true
-
-# Log level for console output
-log-level=DEBUG
-
-# Log level for file output
-log-level-file=WARN
-
-# Network (signet or regtest)
-network=%s
-
-# RPC server address
-rpc-addr=%s
-
-# P2P networking address
-net-addr=%s
-
-# Mainchain (Enforcer) gRPC connection URL
-mainchain-grpc-url=%s
-`, network, ports["rpc-addr"], ports["net-addr"], ports["mainchain-grpc-url"])
-}
-
-// getZSideNetworkPorts returns network-specific port mappings for ZSide.
-func getZSideNetworkPorts(network string) map[string]string {
-	switch network {
-	case "regtest":
-		return map[string]string{
-			"rpc-addr":           "127.0.0.1:16098",
-			"net-addr":           "0.0.0.0:14098",
-			"mainchain-grpc-url": "http://localhost:50051",
-		}
-	case "mainnet":
-		return map[string]string{
-			"rpc-addr":           "127.0.0.1:26098",
-			"net-addr":           "0.0.0.0:24098",
-			"mainchain-grpc-url": "http://localhost:50051",
-		}
-	default: // signet
-		return map[string]string{
-			"rpc-addr":           "127.0.0.1:6098",
-			"net-addr":           "0.0.0.0:4098",
-			"mainchain-grpc-url": "http://localhost:50051",
-		}
-	}
-}
-
-// GetCliArgs converts current config to CLI args.
-func (m *ZSideConfManager) GetCliArgs() []string {
-	var args []string
-	if m.Config == nil {
-		return args
-	}
-
-	for key, value := range m.Config.Settings {
-		if key == "network" {
-			continue
-		}
-		switch value {
-		case "true":
-			args = append(args, fmt.Sprintf("--%s", key))
-		case "false":
-			continue
-		default:
-			if value != "" {
-				args = append(args, fmt.Sprintf("--%s=%s", key, value))
-			}
-		}
-	}
-
-	return args
-}
-
-// GetNetwork returns the current network from config.
-func (m *ZSideConfManager) GetNetwork() string {
-	if m.Config == nil {
-		return "signet"
-	}
-	n := m.Config.GetSetting("network")
-	if n == "" {
-		return "signet"
-	}
-	return n
-}
-
-// GetCurrentConfigContent returns config as string.
-func (m *ZSideConfManager) GetCurrentConfigContent() string {
-	if m.Config == nil {
-		return m.GetDefaultConfig()
-	}
-	return m.Config.Serialize()
-}
-
-// WriteConfig writes raw config content to file.
-func (m *ZSideConfManager) WriteConfig(content string) error {
-	m.Config = ParseGenericAppConfig(content)
-
-	confPath := m.getConfigPath()
-	if err := os.MkdirAll(filepath.Dir(confPath), 0755); err != nil {
-		return fmt.Errorf("create dir: %w", err)
-	}
-	if err := os.WriteFile(confPath, []byte(content), 0644); err != nil {
-		return fmt.Errorf("write config: %w", err)
-	}
-
-	m.log.Info().Str("path", confPath).Msg("saved zside config")
-	return nil
-}
-
-// SyncNetworkFromBitcoinConf syncs network setting from Bitcoin config.
-func (m *ZSideConfManager) SyncNetworkFromBitcoinConf() error {
-	if m.Config == nil || m.bitcoinConf == nil {
-		return nil
-	}
-
-	var sidechainNetwork string
-	switch m.bitcoinConf.Network {
-	case NetworkRegtest:
-		sidechainNetwork = "regtest"
-	case NetworkECash:
-		sidechainNetwork = "mainnet"
-	default:
-		sidechainNetwork = "signet"
-	}
-
-	currentNetwork := m.Config.GetSetting("network")
-	if currentNetwork != sidechainNetwork {
-		m.Config.SetSetting("network", sidechainNetwork)
-		ports := getZSideNetworkPorts(sidechainNetwork)
-		for key, value := range ports {
-			m.Config.SetSetting(key, value)
-		}
-		return m.saveConfig()
-	}
-
-	return nil
-}
-
-func (m *ZSideConfManager) saveConfig() error {
-	if m.Config == nil {
-		return nil
-	}
-	confPath := m.getConfigPath()
-	if err := os.MkdirAll(filepath.Dir(confPath), 0755); err != nil {
-		return err
-	}
-	if err := os.WriteFile(confPath, []byte(m.Config.Serialize()), 0644); err != nil {
-		return fmt.Errorf("save zside config: %w", err)
-	}
-	m.log.Info().Str("path", confPath).Msg("saved zside config")
-	return nil
-}
-
-func (m *ZSideConfManager) getConfigPath() string {
-	return filepath.Join(ZSideDirs.RootDir(), zsideConfFilename)
-}
-
-// StartWatching watches the zside config directory for changes.
-func (m *ZSideConfManager) StartWatching() error {
-	watcher, err := fsnotify.NewWatcher()
-	if err != nil {
-		return fmt.Errorf("create watcher: %w", err)
-	}
-
-	confDir := filepath.Dir(m.getConfigPath())
-	if err := os.MkdirAll(confDir, 0755); err != nil {
-		_ = watcher.Close()
-		return fmt.Errorf("create watch dir: %w", err)
-	}
-
-	if err := watcher.Add(confDir); err != nil {
-		_ = watcher.Close()
-		return fmt.Errorf("watch dir: %w", err)
-	}
-
-	m.watcher = watcher
-	m.watchDone = make(chan struct{})
-
-	go m.watchLoop()
-
-	m.log.Debug().Str("dir", confDir).Msg("zside config file watching enabled")
-	return nil
-}
-
-func (m *ZSideConfManager) StopWatching() {
-	if m.watcher != nil {
-		_ = m.watcher.Close()
-	}
-	if m.watchDone != nil {
-		<-m.watchDone
-	}
-}
-
-func (m *ZSideConfManager) watchLoop() {
-	defer close(m.watchDone)
-
-	var debounce *time.Timer
-	var mu sync.Mutex
-
-	for {
-		select {
-		case event, ok := <-m.watcher.Events:
-			if !ok {
-				return
-			}
-			if !strings.HasSuffix(event.Name, zsideConfFilename) {
-				continue
-			}
-			if event.Op&(fsnotify.Write|fsnotify.Create) == 0 {
-				continue
-			}
-
-			mu.Lock()
-			if debounce != nil {
-				debounce.Stop()
-			}
-			debounce = time.AfterFunc(500*time.Millisecond, func() {
-				m.reloadConfigFromFileSystem()
-			})
-			mu.Unlock()
-
-		case err, ok := <-m.watcher.Errors:
-			if !ok {
-				return
-			}
-			m.log.Error().Err(err).Msg("zside config watcher error")
-		}
-	}
-}
-
-func (m *ZSideConfManager) reloadConfigFromFileSystem() {
-	m.log.Info().Msg("reloading zside config due to file system change")
-
-	confPath := m.getConfigPath()
-	data, err := os.ReadFile(confPath)
-	if err != nil {
-		m.log.Error().Err(err).Msg("failed to read zside config")
-		return
-	}
-
-	newConfig := ParseGenericAppConfig(string(data))
-	if m.Config != nil && m.Config.Serialize() == newConfig.Serialize() {
-		return
-	}
-
-	m.Config = newConfig
-}
